test(cli): reuse shared binary builder
Move the verdict-exit binary regressions into the external test package so
they can share the existing once-built CLI binary.
